* mcc/HandleInput.c, mcc/SpellChecker.c: cast all characters passed to any locale.library function to ULONG to avoid any possible sign extension which might cause the called function to return wrong results.

+++ mcc/HandleInput.c
@@ -487,7 +487,7 @@ void Key_Normal(struct InstData *data, char key)
 
   // check if the user wants to do a direct spell checking while
   // writing some text.
-  if(key != '-' && IsAlpha(data->mylocale, key) == FALSE)
+  if(key != '-' && IsAlpha(data->mylocale, (ULONG)key) == FALSE)
   {
     // only perform a spell check if it has been enabled
     if(data->TypeAndSpell == TRUE)
